Also check chroot_dir when copying install log. Useful when only doing disk operations.

+++ modules/install_steps.sh
@@ -548,7 +548,7 @@ starting_cleanup() {
 }
 
 finishing_cleanup() {
-    if [ -f ${logfile} ]; then
+    if [ -f ${logfile} ] && [ -d ${chroot_dir} ]; then
         spawn "cp ${logfile} ${chroot_dir}/root/$(basename ${logfile})" || warn "could not copy install logfile into chroot"
     fi
